move edit all to levelbuilder folder
